modular controller failed :(
